SpreadsheetCellReferenceOrLabelName deleted
- SpreadsheetCellReference & SpreadsheetLabelName extend SpreadsheetExpressionReference
