--- src/main/java/walkingkooka/spreadsheet/reference/SpreadsheetCellReferenceOrLabelName.java
@@ -1,63 +1,0 @@
-/*
- * Copyright 2019 Miroslav Pokorny (github.com/mP1)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package walkingkooka.spreadsheet.reference;
-
-import walkingkooka.spreadsheet.SpreadsheetViewport;
-
-import java.util.EnumSet;
-import java.util.Set;
-
-/**
- * Base class shared by {@link SpreadsheetCellReference} and {@link SpreadsheetLabelName}.
- * This type is necessary to avoid J2clTranspiler failures because of what appears to be a failure of the JDT not being
- * able to handle type parameters with multiple bounds.
- */
-abstract public class SpreadsheetCellReferenceOrLabelName extends SpreadsheetExpressionReference {
-
-    /**
-     * Package private to limit sub classing.
-     */
-    SpreadsheetCellReferenceOrLabelName() {
-        super();
-    }
-
-    @Override
-    public SpreadsheetSelection simplify() {
-        return this;
-    }
-
-    @Override
-    final Set<SpreadsheetViewportSelectionAnchor> anchors() {
-        return ANCHORS;
-    }
-
-    private final Set<SpreadsheetViewportSelectionAnchor> ANCHORS = EnumSet.of(SpreadsheetViewportSelectionAnchor.NONE);
-
-    // SpreadsheetViewport..............................................................................................
-
-    /**
-     * Creates a {@link SpreadsheetViewport} using this as the top/left.
-     */
-    public final SpreadsheetViewport viewport(final double width,
-                                              final double height) {
-        return SpreadsheetViewport.with(this, width, height);
-    }
-
-    @Override
-    abstract public SpreadsheetCellReferenceOrLabelName toRelative();
-}
